Split date/time related tests into own file
